[mipi] Refactor constants and functions
