+++ esphome/components/const/__init__.py
@@ -3,8 +3,12 @@
 CODEOWNERS = ["@esphome/core"]
 
 CONF_BYTE_ORDER = "byte_order"
+BYTE_ORDER_LITTLE = "little_endian"
+BYTE_ORDER_BIG = "big_endian"
+
 CONF_COLOR_DEPTH = "color_depth"
 CONF_DRAW_ROUNDING = "draw_rounding"
 CONF_ON_RECEIVE = "on_receive"
 CONF_ON_STATE_CHANGE = "on_state_change"
 CONF_REQUEST_HEADERS = "request_headers"
+CONF_USE_PSRAM = "use_psram"
