+++ esphome/components/mipi_spi/__init__.py
@@ -3,11 +3,4 @@ CODEOWNERS = ["@clydebarrow"]
 DOMAIN = "mipi_spi"
 
 CONF_SPI_16 = "spi_16"
-CONF_PIXEL_MODE = "pixel_mode"
 CONF_BUS_MODE = "bus_mode"
-CONF_USE_AXIS_FLIPS = "use_axis_flips"
-CONF_NATIVE_WIDTH = "native_width"
-CONF_NATIVE_HEIGHT = "native_height"
-
-MODE_RGB = "RGB"
-MODE_BGR = "BGR"
