--- esphome/components/mipi_spi/models/commands.py
@@ -1,82 +1,0 @@
-# MIPI DBI commands
-
-NOP = 0x00
-SWRESET = 0x01
-RDDID = 0x04
-RDDST = 0x09
-RDMODE = 0x0A
-RDMADCTL = 0x0B
-RDPIXFMT = 0x0C
-RDIMGFMT = 0x0D
-RDSELFDIAG = 0x0F
-SLEEP_IN = 0x10
-SLPIN = 0x10
-SLEEP_OUT = 0x11
-SLPOUT = 0x11
-PTLON = 0x12
-NORON = 0x13
-INVERT_OFF = 0x20
-INVOFF = 0x20
-INVERT_ON = 0x21
-INVON = 0x21
-ALL_ON = 0x23
-WRAM = 0x24
-GAMMASET = 0x26
-MIPI = 0x26
-DISPOFF = 0x28
-DISPON = 0x29
-CASET = 0x2A
-PASET = 0x2B
-RASET = 0x2B
-RAMWR = 0x2C
-WDATA = 0x2C
-RAMRD = 0x2E
-PTLAR = 0x30
-VSCRDEF = 0x33
-TEON = 0x35
-MADCTL = 0x36
-MADCTL_CMD = 0x36
-VSCRSADD = 0x37
-IDMOFF = 0x38
-IDMON = 0x39
-COLMOD = 0x3A
-PIXFMT = 0x3A
-GETSCANLINE = 0x45
-BRIGHTNESS = 0x51
-WRDISBV = 0x51
-RDDISBV = 0x52
-WRCTRLD = 0x53
-SWIRE1 = 0x5A
-SWIRE2 = 0x5B
-IFMODE = 0xB0
-FRMCTR1 = 0xB1
-FRMCTR2 = 0xB2
-FRMCTR3 = 0xB3
-INVCTR = 0xB4
-DFUNCTR = 0xB6
-ETMOD = 0xB7
-PWCTR1 = 0xC0
-PWCTR2 = 0xC1
-PWCTR3 = 0xC2
-PWCTR4 = 0xC3
-PWCTR5 = 0xC4
-VMCTR1 = 0xC5
-IFCTR = 0xC6
-VMCTR2 = 0xC7
-GMCTR = 0xC8
-SETEXTC = 0xC8
-PWSET = 0xD0
-VMCTR = 0xD1
-PWSETN = 0xD2
-RDID4 = 0xD3
-RDINDEX = 0xD9
-RDID1 = 0xDA
-RDID2 = 0xDB
-RDID3 = 0xDC
-RDIDX = 0xDD
-GMCTRP1 = 0xE0
-GMCTRN1 = 0xE1
-CSCON = 0xF0
-PWCTR6 = 0xF6
-ADJCTL3 = 0xF7
-PAGESEL = 0xFE
