+++ esphome/components/rpi_dpi_rgb/display.py
@@ -2,6 +2,18 @@ from esphome import pins
 import esphome.codegen as cg
 from esphome.components import display
 from esphome.components.esp32 import const, only_on_variant
+from esphome.components.mipi import (
+    CONF_DE_PIN,
+    CONF_HSYNC_BACK_PORCH,
+    CONF_HSYNC_FRONT_PORCH,
+    CONF_HSYNC_PULSE_WIDTH,
+    CONF_PCLK_FREQUENCY,
+    CONF_PCLK_INVERTED,
+    CONF_PCLK_PIN,
+    CONF_VSYNC_BACK_PORCH,
+    CONF_VSYNC_FRONT_PORCH,
+    CONF_VSYNC_PULSE_WIDTH,
+)
 import esphome.config_validation as cv
 from esphome.const import (
     CONF_BLUE,
@@ -26,18 +38,6 @@ from esphome.const import (
 )
 
 DEPENDENCIES = ["esp32"]
-
-CONF_DE_PIN = "de_pin"
-CONF_PCLK_PIN = "pclk_pin"
-
-CONF_HSYNC_FRONT_PORCH = "hsync_front_porch"
-CONF_HSYNC_PULSE_WIDTH = "hsync_pulse_width"
-CONF_HSYNC_BACK_PORCH = "hsync_back_porch"
-CONF_VSYNC_FRONT_PORCH = "vsync_front_porch"
-CONF_VSYNC_PULSE_WIDTH = "vsync_pulse_width"
-CONF_VSYNC_BACK_PORCH = "vsync_back_porch"
-CONF_PCLK_FREQUENCY = "pclk_frequency"
-CONF_PCLK_INVERTED = "pclk_inverted"
 
 rpi_dpi_rgb_ns = cg.esphome_ns.namespace("rpi_dpi_rgb")
 RPI_DPI_RGB = rpi_dpi_rgb_ns.class_("RpiDpiRgb", display.Display, cg.Component)
